Update spf4.js
Added the parameter keyHeight: allows increasing the distance between the key label and the category labels. 
Added the parameter instTop: allows setting the top location of the block's instruction. 
Changed the default keys because the previous default keys did not work well with an AZERTY keyboard.
Changed the instructions of the SPF because to refer to middle and index fingers rather than pinky and index fingers.